Chobits Recap Episodes (#81)
The recap episodes on anidb are included in the normal runtime order but listed as special episodes  for tvdb. This Properly maps the anidb specials with their tvdb counterparts, while also properly setting the offset between anidb and tvdb to account for the recap episodes, then mapping the tvdb recap specials to the anidb eps.